add manual remap of cols in CSV dump
This helps with readability and usability
(hopefully) across different software.

As an example, only underscore and dot
are used in colnames in CSV now, and all
are lower case, which should help.

Also, the ordering now makes more
sense for our project.

Ultimately, this perhaps will be
replaced with a generation from
knowledge graph (or not).