MAJOR fix [wip]: graph saving
Previously many triples were lost
due to the incomplete transition
to vector row processing.

Now this works and p_gf is better
saved now, but triples are lost
after re-read from serialization,
which is an issue but not so much
now because we have not been using
graph as main storage so far.

At the very least, new API retrievals
seem to be saving in graph now.
Will have to debug and fix further.

Also, there are some duplicate object
statements in master graph, which
I am not quite sure where come from.