pilot sampler: nullable affiliation sort
Implements adding affiliation priority
column at all times when affiliation_sort
is not None, but setting it to False
explicitly disables the actual sorting.

I should also change this ideally
in the main 2025-08-19 sampler
but will not touch it for now as
this does not add any material
change and no new samples
will be drawn soon, particularly
not with affiliation_sort set to False.